git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/branches/HADOOP-10388@1619012 13f79535-47bb-0310-9956-ffa450edef68

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/fs/shell/MoveCommands.java
@@ -45,7 +45,7 @@ class MoveCommands {
     public static final String NAME = "moveFromLocal";
     public static final String USAGE = "<localsrc> ... <dst>";
     public static final String DESCRIPTION = 
-      "Same as -put, except that the source is\n" +
+      "Same as -put, except that the source is " +
       "deleted after it's copied.";
 
     @Override
@@ -87,8 +87,8 @@ class MoveCommands {
     public static final String NAME = "mv";
     public static final String USAGE = "<src> ... <dst>";
     public static final String DESCRIPTION = 
-      "Move files that match the specified file pattern <src>\n" +
-      "to a destination <dst>.  When moving multiple files, the\n" +
+      "Move files that match the specified file pattern <src> " +
+      "to a destination <dst>.  When moving multiple files, the " +
       "destination must be a directory.";
 
     @Override
@@ -103,6 +103,9 @@ class MoveCommands {
       if (!src.fs.getUri().equals(target.fs.getUri())) {
         throw new PathIOException(src.toString(),
             "Does not match target filesystem");
+      }
+      if (target.exists) {
+        throw new PathExistsException(target.toString());
       }
       if (!target.fs.rename(src.path, target.path)) {
         // we have no way to know the actual error...
